Add files via upload
I tried both the algorithms and found that Decision Tree is performing better and giving 81% accuracy and Support Vector Machine algorithm gave 71% accuracy. In the Desicion Tree my depth was 4 and testing size was 0.05% and in Support Vector Machine testing size was 0.05% and when i was increasing or decreasing the testing size accuracy was decreasing. Overall, I found Decision Tree is performing better in this dataset.